Change edit mode navigation to be tabbed
